remove zoom to label
